refactor: moved page components out of feed
