+++ .dockerignore
@@ -1,4 +1,3 @@
-**/*.md
 **/.gitkeep
 **/.vscode
 .all-contributorsrc
